disco_service/requirements.txt
django-haystack
pysolr
celery
httplib2
lxml
pyorient
django-celery

